# contracts/openapi.yml
openapi: 3.0.3
info:
  title: Kodama Brain API
  version: "0.1.0"
  description: |
    REST surface for managing templates and template versions.
    All endpoints are served by the Brain service.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login
      description: Exchange credentials for a bearer token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/instances:
    get:
      tags: [Instances]
      summary: List instances
      description: Returns all instances with their template layers.
      responses:
        "200":
          description: Instances ordered by persistence default.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Instance"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Instances]
      summary: Create instance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInstanceRequest"
      responses:
        "201":
          description: Instance created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instance"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/instances/{id}:
    parameters:
      - $ref: "#/components/parameters/InstanceId"
    get:
      tags: [Instances]
      summary: Get instance
      responses:
        "200":
          description: Instance found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Instance"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/nodes:
    get:
      tags: [Nodes]
      summary: List nodes
      responses:
        "200":
          description: Nodes ordered by name.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Node"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/nodes/register:
    post:
      tags: [Nodes]
      summary: Register node
      description: Insert or refresh node metadata.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NodeRegistrationRequest"
      responses:
        "200":
          description: Node registered or updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeRegistrationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/nodes/{nodeId}/heartbeat:
    parameters:
      - $ref: "#/components/parameters/NodeId"
    post:
      tags: [Nodes]
      summary: Send heartbeat
      description: Update node status and slot usage.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NodeHeartbeatRequest"
      responses:
        "200":
          description: Heartbeat processed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Node"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/nodes/{nodeId}/instances/{instanceId}/prepared:
    parameters:
      - $ref: "#/components/parameters/NodeId"
      - $ref: "#/components/parameters/NodeInstanceId"
    post:
      tags: [Nodes]
      summary: Report instance prepared
      description: Callback from node when an instance finishes preparation.
      responses:
        "200":
          description: Callback accepted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/nodes/{nodeId}/instances/{instanceId}/running:
    parameters:
      - $ref: "#/components/parameters/NodeId"
      - $ref: "#/components/parameters/NodeInstanceId"
    post:
      tags: [Nodes]
      summary: Report instance running
      description: Callback from node when an instance is running.
      responses:
        "200":
          description: Callback accepted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/nodes/{nodeId}/instances/{instanceId}/stopped:
    parameters:
      - $ref: "#/components/parameters/NodeId"
      - $ref: "#/components/parameters/NodeInstanceId"
    post:
      tags: [Nodes]
      summary: Report instance stopped
      description: Callback from node when an instance stops.
      responses:
        "200":
          description: Callback accepted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/nodes/{nodeId}/instances/{instanceId}/destroyed:
    parameters:
      - $ref: "#/components/parameters/NodeId"
      - $ref: "#/components/parameters/NodeInstanceId"
    post:
      tags: [Nodes]
      summary: Report instance destroyed
      description: Callback from node when an instance is destroyed.
      responses:
        "200":
          description: Callback accepted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/nodes/{nodeId}/instances/{instanceId}/failed:
    parameters:
      - $ref: "#/components/parameters/NodeId"
      - $ref: "#/components/parameters/NodeInstanceId"
    post:
      tags: [Nodes]
      summary: Report instance failed
      description: Callback from node when an instance reports failure.
      responses:
        "200":
          description: Callback accepted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/templates:
    get:
      tags: [Templates]
      summary: List templates
      description: Returns all templates.
      responses:
        "200":
          description: Templates ordered by creation time.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Templates]
      summary: Create template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplateRequest"
      responses:
        "201":
          description: Template created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/templates/{id}:
    parameters:
      - $ref: "#/components/parameters/TemplateId"
    get:
      tags: [Templates]
      summary: Get template
      responses:
        "200":
          description: Template found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Template"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/templates/{id}/versions:
    parameters:
      - $ref: "#/components/parameters/TemplateId"
    get:
      tags: [Template Versions]
      summary: List template versions
      description: Returns versions for the template, newest first.
      responses:
        "200":
          description: Template versions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TemplateVersion"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [Template Versions]
      summary: Add template version
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTemplateVersionRequest"
      responses:
        "201":
          description: Version created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TemplateVersion"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    InstanceId:
      name: id
      in: path
      required: true
      description: Instance identifier.
      schema:
        type: string
        format: uuid
    NodeId:
      name: nodeId
      in: path
      required: true
      description: Node identifier.
      schema:
        type: string
        format: uuid
    NodeInstanceId:
      name: instanceId
      in: path
      required: true
      description: Instance identifier.
      schema:
        type: string
        format: uuid
    TemplateId:
      name: id
      in: path
      required: true
      description: Template identifier.
      schema:
        type: string
        format: uuid
  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
    LoginResponse:
      type: object
      required: [accessToken, tokenType, expiresAt, roles]
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresAt:
          type: string
          format: date-time
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
    Node:
      type: object
      required: [id, name, region, status, devMode, capacitySlots, usedSlots, lastHeartbeatAt, nodeVersion]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        region:
          type: string
        status:
          $ref: "#/components/schemas/NodeStatus"
        devMode:
          type: boolean
        capacitySlots:
          type: integer
          format: int32
        usedSlots:
          type: integer
          format: int32
        lastHeartbeatAt:
          type: string
          format: date-time
        nodeVersion:
          type: string
        tags:
          type: string
          nullable: true
        baseUrl:
          type: string
          format: uri
          nullable: true
    NodeRegistrationRequest:
      type: object
      required: [name, region, capacitySlots, nodeVersion]
      properties:
        name:
          type: string
        region:
          type: string
        capacitySlots:
          type: integer
          format: int32
          minimum: 1
        nodeVersion:
          type: string
        devMode:
          type: boolean
          default: false
        tags:
          type: string
          nullable: true
        baseUrl:
          type: string
          format: uri
          nullable: true
        status:
          $ref: "#/components/schemas/NodeStatus"
          description: Optional override; defaults to ONLINE.
    NodeHeartbeatRequest:
      type: object
      required: [status, usedSlots]
      properties:
        status:
          $ref: "#/components/schemas/NodeStatus"
        usedSlots:
          type: integer
          format: int32
          minimum: 0
    NodeRegistrationResponse:
      type: object
      required: [nodeId, heartbeatIntervalSeconds]
      properties:
        nodeId:
          type: string
          format: uuid
        heartbeatIntervalSeconds:
          type: integer
          format: int32
          description: Preferred heartbeat cadence communicated to the node.
    NodeStatus:
      type: string
      enum: [ONLINE, OFFLINE, UNKNOWN]
    Template:
      type: object
      required: [id, name, description, type, createdAt, createdBy]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        type:
          $ref: "#/components/schemas/TemplateType"
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
    TemplateVersion:
      type: object
      required: [id, templateId, version, checksum, s3Key, createdAt]
      properties:
        id:
          type: string
          format: uuid
        templateId:
          type: string
          format: uuid
        version:
          type: string
        checksum:
          type: string
        s3Key:
          type: string
        metadataJson:
          type: string
          nullable: true
          description: Optional metadata blob persisted with the version.
        createdAt:
          type: string
          format: date-time
    CreateTemplateRequest:
      type: object
      required: [name, description, type, createdBy]
      properties:
        name:
          type: string
        description:
          type: string
        type:
          $ref: "#/components/schemas/TemplateType"
        createdBy:
          type: string
          format: uuid
    CreateTemplateVersionRequest:
      type: object
      required: [version, checksum, s3Key]
      properties:
        version:
          type: string
        checksum:
          type: string
          description: Hash of the uploaded object (client-supplied).
        s3Key:
          type: string
          description: S3 key where the template tarball is stored.
        metadataJson:
          type: string
          nullable: true
          description: Optional metadata stored with the version.
    TemplateType:
      type: string
      enum: [CUSTOM]
    Role:
      type: string
      enum: [ADMIN, OPERATOR, VIEWER]
    Instance:
      type: object
      required:
        [id, name, displayName, state, createdAt, updatedAt, templateLayers]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        displayName:
          type: string
        state:
          $ref: "#/components/schemas/InstanceState"
        nodeId:
          type: string
          format: uuid
          nullable: true
        requestedBy:
          type: string
          format: uuid
          nullable: true
        region:
          type: string
          nullable: true
          description: Optional node preference for scheduling.
        tags:
          type: string
          nullable: true
          description: Optional node preference; comma-separated tags.
        devModeAllowed:
          type: boolean
          nullable: true
          description: Optional node preference; when false, dev-mode nodes are excluded.
        portsJson:
          type: string
          nullable: true
        variablesJson:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
        stoppedAt:
          type: string
          format: date-time
          nullable: true
        failureReason:
          type: string
          nullable: true
        templateLayers:
          type: array
          items:
            $ref: "#/components/schemas/InstanceTemplateLayer"
    InstanceTemplateLayer:
      type: object
      required: [id, templateVersionId, orderIndex]
      properties:
        id:
          type: string
          format: uuid
        templateVersionId:
          type: string
          format: uuid
        orderIndex:
          type: integer
          format: int32
          minimum: 0
    CreateInstanceRequest:
      type: object
      required: [name, displayName, templateLayers]
      properties:
        name:
          type: string
        displayName:
          type: string
        requestedBy:
          type: string
          format: uuid
          nullable: true
        nodeId:
          type: string
          format: uuid
          nullable: true
        region:
          type: string
          nullable: true
        tags:
          type: string
          nullable: true
          description: Optional node preference; comma-separated tags.
        devModeAllowed:
          type: boolean
          nullable: true
        templateLayers:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/InstanceTemplateLayerRequest"
        variables:
          type: object
          nullable: true
          additionalProperties:
            type: string
          description: Variables map serialized into `variablesJson`. Mutually exclusive with `variablesJson`.
        variablesJson:
          type: string
          nullable: true
          description: Raw JSON string persisted verbatim. Mutually exclusive with `variables`.
        portsJson:
          type: string
          nullable: true
      not:
        required: [variables, variablesJson]
    InstanceTemplateLayerRequest:
      type: object
      description: Either `templateId` or `templateVersionId` is required. If `orderIndex` is omitted, list order is used.
      required: []
      properties:
        templateVersionId:
          type: string
          format: uuid
          description: Reference a specific template version.
        templateId:
          type: string
          format: uuid
          description: Reference a template; latest version is used when no version id is provided.
        orderIndex:
          type: integer
          format: int32
          minimum: 0
          description: Optional explicit order; defaults to list order position.
    InstanceState:
      type: string
      enum:
        [REQUESTED, PREPARING, PREPARED, STARTING, RUNNING, STOPPING, STOPPED, DESTROYED, FAILED]
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
      required: [status, error, message]
  responses:
    BadRequest:
      description: Validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or invalid credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated but not allowed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource already exists.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"